requirements.txt
fastapi>=0.68.0,<0.69.0
uvicorn>=0.15.0,<0.16.0
sqlalchemy>=1.4.0,<1.5.0
alembic>=1.12.1
psycopg2-binary==2.9.9
python-dotenv>=0.19.0,<0.20.0
celery>=5.3.6
redis>=5.0.1
scrapy>=2.12.0
trafilatura>=1.6.1
langchain>=0.1.0
python-dateutil>=2.8.2
beautifulsoup4>=4.12.2
lxml>=4.9.3
aiohttp>=3.8.0,<3.9.0
python-jose[cryptography]>=3.3.0,<3.4.0
tweepy==4.14.0
anthropic==0.8.1
openai==1.6.1
python-multipart>=0.0.5,<0.1.0
passlib[bcrypt]>=1.7.4,<1.8.0
bcrypt==4.1.2
linkedin-api>=2.0.3
pytest>=7.4.3
httpx>=0.25.1
pydantic>=1.9.0,<2.0.0
asyncpg>=0.29.0
cryptography>=37.0.0
fastapi-sessions==0.3.2
itsdangerous==2.1.2
